zynq_gem: Remove writing to pad skew reigsters.
Adjusting the pad skew in the phy turns out to not be needed. Ethernet
works fine without these adjustments.

The root cause of ethernet not working was calling zynq_gem_initialize()
with the 'emio' argument of 0 when a 1 is needed (we use the GMII to
RGMII ip core which uses EMIO). This caused RCLK and CLK to get changed
when they should not have been.